Improve negative tests
