+++ misc/debug-adapter/modules/debug-adapter-core/src/main/java/org/ballerinalang/debugadapter/evaluation/DebugExpressionEvaluator.java
@@ -66,7 +66,7 @@ public class DebugExpressionEvaluator extends Evaluator {
 
             EvaluatorBuilder evaluatorBuilder = new EvaluatorBuilder(evaluationContext);
             Evaluator evaluator = evaluatorBuilder.build(parsedExpression);
-            return new BExpressionValue(context, evaluator.evaluate().getJdiValue());
+            return evaluator.evaluate();
         } catch (EvaluationException e) {
             throw e;
         } catch (Exception e) {
