+++ tests/jballerina-debugger-integration-test/src/test/java/org/ballerinalang/debugger/test/adapter/evaluation/EvaluationExceptionKind.java
@@ -34,6 +34,7 @@ public enum EvaluationExceptionKind {
     FIELD_NOT_FOUND("Undefined field '%s' in `%s`"),
     VARIABLE_NOT_FOUND("Undefined variable '%s'"),
     QUALIFIED_VARIABLE_RESOLVING_FAILED("Failed to resolve qualified variable: '%s:%s'"),
+    QUALIFIED_REFERENCE_NOT_FOUND("Failed to resolve qualified name reference: '%s:%s'"),
     HELPER_UTIL_NOT_FOUND("Failed to load the evaluation helper util method: '%s'"),
     FUNCTION_EXECUTION_ERROR("Error occurred when executing method: '%s'"),
     INDEX_OUT_OF_RANGE_ERROR("%s index out of range: index=%d, size=%d"),
@@ -49,6 +50,7 @@ public enum EvaluationExceptionKind {
     MISSING_MESSAGE_IN_ERROR("Missing error message in error constructor"),
     REST_ARG_IN_ERROR("Rest args are not allowed in error constructor"),
     ADDITIONAL_ARG_IN_ERROR("Additional positional arg in error constructor"),
+
     CANNOT_INFER_PARAM_TYPE("Can not infer type for parameter '%s' in function '%s'. Consider passing" +
             "the parameter value explicitly as an argument"),
     VARIABLE_EXECUTION_ERROR("Internal error occurred when processing variable: '%s'"),
