+++ tests/jballerina-debugger-integration-test/src/test/java/org/ballerinalang/debugger/test/adapter/evaluation/SingleFileEvaluationNegativeTest.java
@@ -91,8 +91,8 @@ public class SingleFileEvaluationNegativeTest extends ExpressionEvaluationNegati
     }
 
     @Override
-    public void simpleNameReferenceEvaluationTest() throws BallerinaTestException {
-        super.simpleNameReferenceEvaluationTest();
+    public void nameReferenceEvaluationTest() throws BallerinaTestException {
+        super.nameReferenceEvaluationTest();
 
         // access constants in undefined modules
         debugTestRunner.assertEvaluationError(context, "other:constant",
