--- tests/jballerina-debugger-integration-test/src/test/resources/project-based-tests/evaluation-tests-1/Dependencies.toml
@@ -1,50 +1,0 @@
-# AUTO-GENERATED FILE. DO NOT MODIFY.
-
-# This file is auto-generated by Ballerina for managing dependency versions.
-# It should not be modified by hand.
-
-[ballerina]
-dependencies-toml-version = "2"
-
-[[package]]
-org = "ballerina"
-name = "jballerina.java"
-version = "0.0.0"
-
-[[package]]
-org = "ballerina"
-name = "lang.float"
-version = "0.0.0"
-dependencies = [
-	{org = "ballerina", name = "jballerina.java"}
-]
-modules = [
-	{org = "ballerina", packageName = "lang.float", moduleName = "lang.float"}
-]
-
-[[package]]
-org = "ballerina"
-name = "lang.int"
-version = "0.0.0"
-dependencies = [
-	{org = "ballerina", name = "jballerina.java"}
-]
-modules = [
-	{org = "ballerina", packageName = "lang.int", moduleName = "lang.int"}
-]
-
-[[package]]
-org = "debug_test_resources"
-name = "evaluation_tests"
-version = "0.1.0"
-dependencies = [
-	{org = "ballerina", name = "lang.float"},
-	{org = "ballerina", name = "lang.int"}
-]
-modules = [
-	{org = "debug_test_resources", packageName = "evaluation_tests", moduleName = "evaluation_tests"},
-	{org = "debug_test_resources", packageName = "evaluation_tests", moduleName = "evaluation_tests.other"},
-	{org = "debug_test_resources", packageName = "evaluation_tests", moduleName = "evaluation_tests.other2"}
-]
-
-
