+++ tests/jballerina-debugger-integration-test/src/test/resources/testng.xml
@@ -31,44 +31,44 @@ under the License.
         </groups>
         <classes>
             <!--Remote Debugging Tests-->
-<!--            <class name="org.ballerinalang.debugger.test.remote.BallerinaRunRemoteDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.remote.BallerinaTestRemoteDebugTest"/>-->
+            <class name="org.ballerinalang.debugger.test.remote.BallerinaRunRemoteDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.remote.BallerinaTestRemoteDebugTest"/>
 
-<!--            &lt;!&ndash;Debug engaging tests for ballerina commands &ndash;&gt;-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.run.SingleBalFileRunDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.ModuleBreakpointTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.run.MultiModuleRunDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.test.MultiModuleTestDebugTest"/>-->
+            <!--Debug engaging tests for ballerina commands -->
+            <class name="org.ballerinalang.debugger.test.adapter.run.SingleBalFileRunDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.ModuleBreakpointTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.run.MultiModuleRunDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.test.MultiModuleTestDebugTest"/>
 
-<!--            &lt;!&ndash;Debug Variables Tests&ndash;&gt;-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.variables.VariableVisibilityTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.variables.VariableQueryTest"/>-->
+            <!--Debug Variables Tests-->
+            <class name="org.ballerinalang.debugger.test.adapter.variables.VariableVisibilityTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.variables.VariableQueryTest"/>
 
             <!--Debugger Expression Evaluation Tests-->
             <class name="org.ballerinalang.debugger.test.adapter.evaluation.PackageEvaluationTest"/>
             <class name="org.ballerinalang.debugger.test.adapter.evaluation.PackageEvaluationNegativeTest"/>
-<!--            <class name="org.ballerinalang.debugger.test.adapter.evaluation.SingleFileEvaluationTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.evaluation.SingleFileEvaluationNegativeTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.evaluation.DependencyEvaluationTest"/>-->
+            <!--            <class name="org.ballerinalang.debugger.test.adapter.evaluation.SingleFileEvaluationTest"/>-->
+            <!--            <class name="org.ballerinalang.debugger.test.adapter.evaluation.SingleFileEvaluationNegativeTest"/>-->
+            <!--            <class name="org.ballerinalang.debugger.test.adapter.evaluation.DependencyEvaluationTest"/>-->
 
             <!--Other tests-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.ControlFlowDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.CallStackDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.LanguageConstructDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.DebugTerminationTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.DebugInstructionTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.LangLibDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.ConditionalBreakpointTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.LogPointTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.run.DebugEngageTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.DebugOutputTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.RecursiveDebugTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.BreakpointVerificationTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.completions.DebugCompletionTest"/>-->
+            <class name="org.ballerinalang.debugger.test.adapter.ControlFlowDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.CallStackDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.LanguageConstructDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.DebugTerminationTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.DebugInstructionTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.LangLibDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.ConditionalBreakpointTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.LogPointTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.run.DebugEngageTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.DebugOutputTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.RecursiveDebugTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.BreakpointVerificationTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.completions.DebugCompletionTest"/>
 
-<!--            &lt;!&ndash;Debugger RunInTerminal Tests&ndash;&gt;-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.runinterminal.SingleFileRunInTerminalTest"/>-->
-<!--            <class name="org.ballerinalang.debugger.test.adapter.runinterminal.ProjectBasedRunInTerminalTest"/>-->
+            <!--Debugger RunInTerminal Tests-->
+            <class name="org.ballerinalang.debugger.test.adapter.runinterminal.SingleFileRunInTerminalTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.runinterminal.ProjectBasedRunInTerminalTest"/>
         </classes>
     </test>
 </suite>
